bunch of changes and fixes for routing
we now remove all resolvers before redirecting so we don't get stuck resolvers (this should fix the bugs with empty pages after a redirect)

watching the $route change was getting set multiple times by the @Watch decorator for some reason. this should fix that.

remove the resolver earlier on in the resolveRoute func to make sure that it's marked as resolved during errors

no longer emitting routeChangeAfter unless it's actually the exact end of the route chain -- we don't trigger it if we're refreshing cache until after the cache is refreshed. we also no longer double call it in the $route watcher