Lab blue hue fix
==== Overview:
This change resolves a minor custom-color selection issue, where choosing dark blues resulted in purplish themes.

==== Details:
Prior to this change, the custom theme color was converted to Lab space for better perceptual brightness clamping. As a side effect of Lab conversion, the output hue was partly distorted from the input hue. In particular, inputting dark blues resulted in a purplish hue in the output color.

This change leaves the Lab conversion process in place, but adds a final step that overwrites the output color's hue with the input color's hue. As a result, if a user picks a custom-color, their theme is guaranteed to have the hue they selected.

==== Notes:
Most of the code added or changed is in the "theme.model.ts" file.
Two functions were added to "color.ts", rgb2hsl() and hsl2rgb(), and these came from the original "theme-blue-fix" branch of the "gamejolt" repo.

==== Files changed:
theme.model.ts
 (In frontend-lib "\components\theme")
 (Or gamejolt "\src\lib\gj-lib-client\components\theme")

color.ts
 (In frontend-lib "\utils")
 (Or gamejolt "\src\lib\gj-lib-client\utils")

Note that these files are both from the "frontend-lib" repo, and are used in, but not directly part of, the "gamejolt" repo.